Updated FilterSelect tests, moved utils flatten tests from filteredData
