+++ src/app/static/src/tests/filteredData/utils.test.ts
@@ -1,7 +1,5 @@
 import {
     colorFunctionFromName,
-    flattenToIdSet,
-    flattenOptions,
     getColor,
     roundToContext
 } from "../../app/store/filteredData/utils";
@@ -82,59 +80,5 @@ describe("FilteredData getters", () => {
 
     it ("round to context does not round value if both values and context are integers", () => {
         expect(roundToContext(5, 10)).toBe(5);
-    });
-
-    it("can flatten options", () => {
-
-        const testData: NestedFilterOption[] = [
-            {
-                id: "1", label: "name1", children: [{
-                    id: "2", label: "nested", children: []
-                }]
-            }
-        ];
-
-        const result = flattenOptions(testData);
-        expect(result["1"]).toStrictEqual(testData[0]);
-        expect(result["2"]).toStrictEqual({id: "2", label: "nested", children: []});
-    });
-
-    it("flatten ids returns set of selected ids", () => {
-
-        const dict = {
-            "1": {
-                id: "1",
-                label: "l1",
-                children:
-                    [{
-                        id: "2",
-                        label: "l2",
-                        children: [{
-                            id: "3",
-                            label: "l3"
-                        }]
-                    }]
-            },
-            "2": {
-                id: "2",
-                label: "l2",
-                children: [{
-                    id: "3",
-                    label: "l3"
-                }]
-            },
-            "3": {
-                id: "3",
-                label: "l3"
-            },
-            "4": {
-                id: "3",
-                label: "l3"
-            }
-        };
-
-
-        const result = flattenToIdSet(["1"], dict);
-        expect(result).toStrictEqual(new Set(["1", "2", "3"]));
     });
 });
