+++ src/app/static/src/tests/utils.test.ts
@@ -1,4 +1,5 @@
-import {addCheckSum, freezer, verifyCheckSum} from "../app/utils";
+import {addCheckSum, freezer, verifyCheckSum, flattenOptions, flattenToIdSet} from "../app/utils";
+import {NestedFilterOption} from "../app/generated";
 
 describe("utils", () => {
 
@@ -58,6 +59,60 @@ describe("utils", () => {
         expect(frozen).toStrictEqual(data);
         expect(Object.isFrozen(frozen[0])).toBe(true);
         expect(Object.isFrozen(frozen[1].child)).toBe(true);
+    });
+
+    it("can flatten options", () => {
+
+        const testData: NestedFilterOption[] = [
+            {
+                id: "1", label: "name1", children: [{
+                    id: "2", label: "nested", children: []
+                }]
+            }
+        ];
+
+        const result = flattenOptions(testData);
+        expect(result["1"]).toStrictEqual(testData[0]);
+        expect(result["2"]).toStrictEqual({id: "2", label: "nested", children: []});
+    });
+
+    it("flatten ids returns set of selected ids", () => {
+
+        const dict = {
+            "1": {
+                id: "1",
+                label: "l1",
+                children:
+                    [{
+                        id: "2",
+                        label: "l2",
+                        children: [{
+                            id: "3",
+                            label: "l3"
+                        }]
+                    }]
+            },
+            "2": {
+                id: "2",
+                label: "l2",
+                children: [{
+                    id: "3",
+                    label: "l3"
+                }]
+            },
+            "3": {
+                id: "3",
+                label: "l3"
+            },
+            "4": {
+                id: "3",
+                label: "l3"
+            }
+        };
+
+
+        const result = flattenToIdSet(["1"], dict);
+        expect(result).toStrictEqual(new Set(["1", "2", "3"]));
     });
 
 });
